navigation improvements, allow blank forms to be selected from a report, re #269
